% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{funcao}
\alias{funcao}
\title{soma um}
\usage{
funcao(x)
}
\arguments{
\item{x}{- valor a ser adicionado um}
}
\value{
x - valor adicinado de um
}
\description{
essa funcao recebe um inteiro e retorna o mesmo numero adicionado de um
}
\examples{

funcao(10)
z<-3
funcao(z)
\dontrun{funcao('blah')}

}

