% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{modelar}
\alias{modelar}
\title{minha funcao que modela}
\usage{
modelar(formula, dados)
}
\arguments{
\item{formula}{minha formula}

\item{dados}{meus dados}
}
\value{
objeto do tipo gam
}
\description{
Ajusta um modelo aditivo generalizado
aos dados
}

