% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{pertenceintervalo}
\alias{pertenceintervalo}
\title{funcao que verifica se um número x
está em um intervalo (a,b)}
\usage{
pertenceintervalo(x, a, b)
}
\arguments{
\item{x}{valor a ser verificado}

\item{a}{minimo do intervalo}

\item{b}{maximo do intervalo}
}
\value{
tabela
}
\description{
funcao que verifica se um número x
está em um intervalo (a,b)
}

