% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{tempoacessodados}
\alias{tempoacessodados}
\title{funcao que compara tempo de acesso a dados}
\usage{
tempoacessodados(dados)
}
\arguments{
\item{dados}{meus dados}
}
\value{
tabela
}
\description{
funcao que compara tempo de acesso a dados
}

