% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes.R
\name{tempoprocessamento}
\alias{tempoprocessamento}
\title{funcao que compara tempo de processamento}
\usage{
tempoprocessamento(dados)
}
\arguments{
\item{dados}{meus dados}
}
\value{
tabela
}
\description{
funcao que compara tempo de processamento
}

